[iOS] Add Home Detect Code
