+++ ios/Classes/SwiftHardwareButtonsPlugin.swift
@@ -1,23 +1,27 @@
 import Flutter
 import UIKit
-import AVFoundation
 
 enum ChannelName {
     static let volume = "flutter.moum.hardware_buttons.volume"
+    static let home = "flutter.moum.hardware_buttons.home"
 }
 
-public class SwiftHardwareButtonsPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
-    
-    private var volumeLevel: Float = 0.0
-    var eventSink: FlutterEventSink?
-    
+public class SwiftHardwareButtonsPlugin: NSObject, FlutterPlugin {
+
     public static func register(with registrar: FlutterPluginRegistrar) {
-        let instance = SwiftHardwareButtonsPlugin()
         
         // VolumeButton
+        let volumeInstance = VolumeButtenStreamHandler()
         let volumeStream = FlutterEventChannel(name: ChannelName.volume,
                                                binaryMessenger: registrar.messenger())
-        volumeStream.setStreamHandler(instance)
+        volumeStream.setStreamHandler(volumeInstance)
+        
+        // HomeButton
+        let homeInstance = HomeButtonStreamHandler()
+        let homeStream = FlutterEventChannel(name: ChannelName.home,
+                                             binaryMessenger: registrar.messenger())
+        homeStream.setStreamHandler(homeInstance)
+        
     }
 
     public func handle(_ call: FlutterMethodCall,
@@ -25,43 +29,4 @@ public class SwiftHardwareButtonsPlugin: NSObject, FlutterPlugin, FlutterStreamH
         result("iOS " + UIDevice.current.systemVersion)
     }
     
-    public func onListen(withArguments arguments: Any?,
-                  eventSink events: @escaping FlutterEventSink) -> FlutterError? {
-        self.eventSink = events
-        let audioSession = AVAudioSession.sharedInstance()
-        do {
-            try audioSession.setActive(true)
-            audioSession.addObserver(self,
-                                     forKeyPath: "outputVolume",
-                                     options: .new,
-                                     context: nil)
-            volumeLevel = audioSession.outputVolume
-        } catch {
-            print("error")
-        }
-        return nil
-    }
-    
-    override public func observeValue(forKeyPath keyPath: String?,
-                                      of object: Any?,
-                                      change: [NSKeyValueChangeKey: Any]?,
-                                      context: UnsafeMutableRawPointer?) {
-        if keyPath == "outputVolume" {
-            let audioSession = AVAudioSession.sharedInstance()
-            if audioSession.outputVolume > volumeLevel {
-                eventSink?(24)
-            }
-            if audioSession.outputVolume < volumeLevel {
-                eventSink?(25)
-            }
-            volumeLevel = audioSession.outputVolume
-        }
-    }
-    
-    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
-        NotificationCenter.default.removeObserver(self,
-                                                  forKeyPath: "outputVolume")
-        eventSink = nil
-        return nil
-    }
 }
